try文を導入してコード量を減らした
